requirements.txt
pysstv
simpleaudio
Pillow>=8.0.0
